fin-rag/app/requirements.txt
# Core app
streamlit>=1.24.0
nest_asyncio
python-dotenv
pysqlite3-binary          # for sqlite on Python ≥3.10
chromadb                  # persistent vector DB
pdfplumber                # PDF extraction (used by llama-parse)

# LangChain & related
langchain
langchain-core            # for langchain_core.messages
langchain-community
langchain-chroma
langchain-google-genai
langchain-groq

# Indexing/parsing
llama-index
llama-parse

# Tokenizers (pulled in by some embeddings/LLM code)
tiktoken
